refactor(transport/http): move swagger/health/cors under transport/server/http
- platform/swagger → transport/server/http/swagger
- platform/health → transport/server/http/health
- transport/server/middleware/cors{.go,_test.go} → transport/server/http/cors/
  （新独立 package cors，与 chain/whitelist 的 kratos middleware 类型分离）
- 删除空的 platform/ 顶层目录
- 更新 transport/server/http/server.go、options.go、server_test.go 的 import

build + cors 26 个 test 通过。
