+++ transport/server/http/cors/cors_test.go
@@ -1,4 +1,4 @@
-package middleware
+package cors
 
 import (
 	"net/http"
